+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - remove legacy version annotation on docker-compose.yml
 
+### Fixed
+- fix Drupal 11 crash by removing legacy field type categories from Template whisperer field
+
 ## [4.0.1] - 2024-03-01
 ### Changed
 - re-enable PHPUnit Symfony Deprecation notice